moving signing interfaces to same interface
