+++ antibytes-publishing/src/main/kotlin/tech/antibytes/gradle/publishing/PublishingApiContract.kt
@@ -76,4 +76,13 @@ interface PublishingApiContract {
     }
 
     interface MavenRepositoryConfiguration : RepositoryConfiguration
+
+    interface MemorySigning {
+        val key: String?
+        val password: String?
+    }
+
+    interface CompleteMemorySigning : MemorySigning {
+        val keyId: String?
+    }
 }
