+++ buildSrc/src/main/kotlin/tech/antibytes/gradle/plugin/script/maven-package.gradle.kts
@@ -27,7 +27,6 @@ import org.gradle.api.publish.maven.MavenPublication
  */
 plugins {
     `maven-publish`
-    `signing`
 }
 
 publishing {
